+++ lib/tr.bash
@@ -104,7 +104,7 @@ function tr_import_export_internal {
         fi
 
         # do the work
-        server_run -d $db -l $lang $extra_opt --i18n-$cmd=$i18n_file --modules=$addon;
+        server_run -- -d $db -l $lang $extra_opt --i18n-$cmd=$i18n_file --modules=$addon;
     done
 }
 
@@ -266,7 +266,7 @@ function tr_load {
     fi
 
     for idb in $(tr_parse_db_name $db); do
-        server_run -d $idb --load-language=$lang --stop-after-init;
+        server_run -- -d $idb --load-language=$lang --stop-after-init;
     done
 }
 
@@ -410,8 +410,8 @@ function tr_translation_rate {
         local trans_tmp_dir=$ODOO_PATH/addons/tmp;
         mkdir -p $trans_tmp_dir;
         local trans_file=$trans_tmp_dir/x-odoo-trans-${tmp_db_name}.po;
-        if server_run -d $tmp_db_name -l $lang --i18n-export=$trans_file --modules=$addons_cs; then
-            if ! server_run -d $tmp_db_name -l $lang --i18n-import=$trans_file --modules=$addons_cs; then
+        if server_run -- -d $tmp_db_name -l $lang --i18n-export=$trans_file --modules=$addons_cs; then
+            if ! server_run -- -d $tmp_db_name -l $lang --i18n-import=$trans_file --modules=$addons_cs; then
                 echoe -e "${REDC}ERROR${NC}: cannot import generated translations";
                 rm $trans_file;
                 res=11;